11450 - Wedding shopping

+++ UVA/11450 - Wedding shopping.cpp
@@ -10,11 +10,11 @@ int main()
         int m,c;
         scanf("%d %d",&m,&c);
 
-        vector<vector<long long> > v;
+        vector<vector<int> > v;
 
         for(int i=0; i<c; ++i){
             int k;
-            vector<long long> temp;
+            vector<int> temp;
             scanf("%d",&k);
             for(int j=0; j<k; ++j){
                 int x;
@@ -24,15 +24,15 @@ int main()
             v.push_back(temp);
         }
 
-        vector<long long> old=v[0];
-        vector<long long> cur,tmp;
+        vector<int> old=v[0];
+        vector<int> cur,tmp;
         for(int i=1; i<c; ++i){
             cur=v[i];
             bitset<205> visited;
             for(int j=0; j<cur.size(); ++j){
-                long long y=cur[j];
+                int y=cur[j];
                 for(int k=0; k<old.size(); ++k){
-                    long long z=old[k]+y;
+                    int z=old[k]+y;
                     if(z<=m and !visited[z]){
                         visited[z]=1;
                         tmp.push_back(z);
@@ -44,11 +44,11 @@ int main()
             tmp.clear();
         }
         if(old.size()){
-            long long mx=0;
+            int mx=0;
             for(int i=0; i<old.size(); ++i){
                 mx=max(mx,old[i]);
             }
-            printf("%lld\n",mx);
+            printf("%d\n",mx);
         }
         else printf("no solution\n");
     }
